Thanh(commet code và tết file test)

See merge request ct25_cpl_java_02/intern_fu_ct_2025_s2_g4/movie!69

+++ backend/src/main/java/com/example/demo/DTO/request/MovieRequest.java
@@ -28,5 +28,5 @@ public class MovieRequest {
     private String actor;
     private String movieProductionCompany;
     private Integer ageLimit;
-    private List<Integer> typeIds;
+    private List<Long> typeIds;
 }
